Add files via upload
These are some of the files that I have been working on. The dataPicker files are for the Time and Date for booking, and CarBooking is a mess, but does show some of the last things I tried, incl. the bare minimum (which I could not get to work either).
I have not touched anyone else's files, so it is presummably only mine that do not work.